requirements.txt
Flask
requests
pytest
pyodbc
mysql-connector-python
azure-storage-blob
python-dotenv